base_authentication/requirements.txt
starlette_exporter
sqlalchemy
pydantic
psycopg2
python-slugify
pyjwt
fastapi
